requirements.txt
alembic==1.15.2
environs==11.2.1
fastapi==0.135.1
opentelemetry-api==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-exporter-otlp==1.32.0
opentelemetry-instrumentation-fastapi==0.53b0
opentelemetry-instrumentation-logging==0.53b0
opentelemetry-instrumentation-sqlalchemy==0.53b0
psycopg2==2.9.10
pytest==8.3.4
pytz==2024.2
structlog==25.2.0
sqlalchemy==2.0.40
uvicorn==0.32.1
wireup==2.9.0
aiokafka==0.12.0
